+++ cardano-api/src/Cardano/Api/Genesis.hs
@@ -1,25 +1,88 @@
 {-# LANGUAGE DerivingStrategies #-}
+{-# LANGUAGE GeneralisedNewtypeDeriving #-}
 {-# LANGUAGE OverloadedStrings #-}
 
 module Cardano.Api.Genesis
   ( ShelleyGenesis(..)
   , shelleyGenesisDefaults
+
+  -- ** Configuration
+  , ByronGenesisConfig
+  , ShelleyGenesisConfig
+  , AlonzoGenesisConfig
+  , ConwayGenesisConfig
+
+  , ShelleyConfig(..)
+  , GenesisHashByron(..)
+  , GenesisHashShelley(..)
+  , GenesisHashAlonzo(..)
+  , GenesisHashConway(..)
+
+  -- ** Files
+  , ByronGenesisFile
+  , ShelleyGenesisFile
+  , AlonzoGenesisFile
+  , ConwayGenesisFile
   ) where
 
+import           Data.ByteString (ByteString)
 import qualified Data.ListMap as ListMap
 import qualified Data.Map.Strict as Map
 import           Data.Maybe (fromMaybe)
+import           Data.Text (Text)
 import qualified Data.Time as Time
 import           Lens.Micro
 
+import qualified Cardano.Crypto.Hash.Blake2b
+import qualified Cardano.Crypto.Hash.Class
+
+import           Cardano.Api.IO
+
+import qualified Cardano.Chain.Genesis
+
+import           Cardano.Ledger.Alonzo.Genesis (AlonzoGenesis (..))
 import           Cardano.Ledger.BaseTypes as Ledger
 import           Cardano.Ledger.Coin (Coin (..))
+import           Cardano.Ledger.Conway.Genesis (ConwayGenesis (..))
 import           Cardano.Ledger.Crypto (StandardCrypto)
 import           Cardano.Ledger.Shelley.Core
 import           Cardano.Ledger.Shelley.Genesis (NominalDiffTimeMicro, ShelleyGenesis (..),
                    emptyGenesisStaking)
 
+import qualified Cardano.Ledger.Shelley.Genesis as Ledger
 
+import qualified Ouroboros.Consensus.Shelley.Eras as Shelley
+
+data ShelleyConfig = ShelleyConfig
+  { scConfig :: !(Ledger.ShelleyGenesis Shelley.StandardCrypto)
+  , scGenesisHash :: !GenesisHashShelley
+  }
+
+newtype GenesisHashByron = GenesisHashByron
+  { unGenesisHashByron :: Text
+  } deriving newtype (Eq, Show)
+
+newtype GenesisHashShelley = GenesisHashShelley
+  { unGenesisHashShelley :: Cardano.Crypto.Hash.Class.Hash Cardano.Crypto.Hash.Blake2b.Blake2b_256 ByteString
+  } deriving newtype (Eq, Show)
+
+newtype GenesisHashAlonzo = GenesisHashAlonzo
+  { unGenesisHashAlonzo :: Cardano.Crypto.Hash.Class.Hash Cardano.Crypto.Hash.Blake2b.Blake2b_256 ByteString
+  } deriving newtype (Eq, Show)
+
+newtype GenesisHashConway = GenesisHashConway
+  { unGenesisHashConway :: Cardano.Crypto.Hash.Class.Hash Cardano.Crypto.Hash.Blake2b.Blake2b_256 ByteString
+  } deriving newtype (Eq, Show)
+
+type ByronGenesisConfig = Cardano.Chain.Genesis.Config
+type ShelleyGenesisConfig = ShelleyConfig
+type AlonzoGenesisConfig = AlonzoGenesis
+type ConwayGenesisConfig = ConwayGenesis Shelley.StandardCrypto
+
+type ByronGenesisFile = File ByronGenesisConfig
+type ShelleyGenesisFile = File ShelleyGenesisConfig
+type AlonzoGenesisFile = File AlonzoGenesisConfig
+type ConwayGenesisFile = File ConwayGenesisConfig
 
 -- | Some reasonable starting defaults for constructing a 'ShelleyGenesis'.
 --
